2023/12/2 前端添加商户未校验参数

+++ src/main/java/com/admin/domain/entity/MerchantEntity.java
@@ -54,9 +54,9 @@ public class MerchantEntity {
     @Column(name = "gold")
     Long gold;
     /**
-     * 渠道
+     * 渠道号
      */
     @Column(name = "channel")
-    Long channel;
+    int channel;
 
 }
